Refactor AGPException package and remove some warnings
